pyproject.toml
[project]
name = "clapslice"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9,<3.10"
dependencies = [
    "av>=15.1.0",
    "cython>=3.2.4",
    "fast-tsp>=0.1.4",
    "ffmpeg>=1.4",
    "jupyterlab>=4.5.6",
    # "madmom>=0.16.1", needs --no-build-isolation
    "kmedoids>=0.5.4",
    "librosa>=0.10.2.post1",
    "madmom>=0.16.1",
    "numpy~=1.22.0",
    "pip>=26.0.1",
    "pysoundfile>=0.9.0.post1",
    "pytorchvideo>=0.1.5",
    "setuptools==81.0.0",
    "sndfile>=0.2.0",
    "sox>=1.5.0",
    "torchaudio>=2.8.0",
    "torchcodec>=0.7.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.3",
    "transformers>=4.57.6",
    "tsp-solver>=0.1",
]
